fix(ai-world): harden routine config overflow and fix wording (2.11B P3)
Addresses the three actionable P3s from 2.11B review (P3 #1/#2 are
forward-looking design notes for 2.11C's execution-ownership
arbitration, not code changes):

- P3 #3: the RoutineWorkStartMs/RoutineWorkEndMs invalid-config
  fallback computed routineDayLengthMs * 2 in int32, which could
  overflow before the /3 division for a sufficiently large (invalid)
  config value. Routed through int64 intermediates instead. The
  default 1200000ms was never at risk - this only hardens the
  fallback path itself.

- P3 #4: several comments/config docs described routine eligibility
  as "a CIVILIAN with both locations set", but RoutineSystem never
  checks AgentType - eligibility is purely HomeLocation/WorkLocation
  both being set. Reworded to say so explicitly, and to stop implying
  AgentType::Civilian is a gate.

+++ src/server/game/AIWorld/AIWorldMgr.cpp
@@ -267,8 +267,13 @@ void AIWorldMgr::Initialize(Trinity::Asio::IoContext& ioContext)
     {
         TC_LOG_WARN("ai.world", "AIWorld.RoutineWorkStartMs/RoutineWorkEndMs ({}, {}) is invalid for day length {}ms, clamping to 1/3 and 2/3 of day length",
             routineWorkStartMs, routineWorkEndMs, routineDayLengthMs);
-        routineWorkStartMs = routineDayLengthMs / 3;
-        routineWorkEndMs = (routineDayLengthMs * 2) / 3;
+
+        // int64 intermediates: routineDayLengthMs * 2 would overflow int32
+        // above roughly INT32_MAX / 2 - unreachable with the default
+        // 1200000ms, but this path is also what a bad config value falls
+        // through to, so it must stay correct at any int32 input.
+        routineWorkStartMs = int32(int64(routineDayLengthMs) / 3);
+        routineWorkEndMs = int32((int64(routineDayLengthMs) * 2) / 3);
     }
 
     _routineScheduleConfig.DayLengthMs = uint32(routineDayLengthMs);
